chore: move option enums from data_preprocessing to constants
